% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcelltype_Seurat.R
\name{addcelltype_Seurat}
\alias{addcelltype_Seurat}
\title{Add cell type to Seurat object}
\usage{
addcelltype_Seurat(obj, cluster_celltype)
}
\arguments{
\item{obj}{A Seurat object with cluster.}

\item{cluster_celltype}{A list, the name is cluster and the entry is cell type.}
}
\value{
A Seurat object with cell type for each cell.
}
\description{
Add cell type to Seurat object
}
\examples{
pbmc_example <- check_cluster(pbmc_test, nfeatures = 100, npcs = 10,
                              dims = 1:10, k.param = 5, resolution = 0.75)
cluster_celltype <- list(Cluster0 = "NK cells", Cluster1 = "Dendritic cells")
pbmc_example <- addcelltype_Seurat(pbmc_example, cluster_celltype)
head(pbmc_example@active.ident)

}
