% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac2rna.R
\name{atac2rna}
\alias{atac2rna}
\title{Convert a peak matrix to a gene activity matrix}
\usage{
atac2rna(
  peaks,
  annotation.file = NULL,
  seq.levels = c(1:22, "X", "Y"),
  include.body = TRUE,
  upstream = 2000,
  downstream = 0
)
}
\arguments{
\item{peaks}{Matrix of peak counts.}

\item{annotation.file}{Path to GTF annotation file.}

\item{seq.levels}{Which seqlevels to keep (corresponds to chromosomes usually).}

\item{include.body}{Include the gene body?}

\item{upstream}{Number of bases upstream to consider.}

\item{downstream}{Number of bases downstream to consider.}
}
\value{
A Seurat object with both ATAC data matrix and gene activity matrix.
}
\description{
This function takes in a peak matrix and an annotation file (gtf)and collapse the peak matrix to a gene activity matrix.
}
\examples{
# Don't run.
# annotation.file <- "../data/Homo_sapiens.GRCh37.82.gtf"
# ATAC_example <- atac2rna(small_ATAC, annotation.file = annotation.file)

}
