% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cluster.R
\name{check_cluster}
\alias{check_cluster}
\title{Checks if the input is already clustered, if not, the function should do it.}
\usage{
check_cluster(obj)
}
\arguments{
\item{obj}{A Seurat object}
}
\value{
If the input is already clustered, return the original object. If not, the function will return a Seurat object will cluster.
}
\description{
Checks if the input is already clustered, if not, the function should do it.
}
\examples{
library(Seurat)
inputMatrix <- matrix(data = rnbinom(n = 1e6, size = 10, prob = .9), nrow = 5000)
rownames(inputMatrix) <- paste0('Gene_', seq_len(nrow(inputMatrix)))
colnames(inputMatrix) <- paste0('Cell_', seq_len(ncol(inputMatrix)))
dta <- CreateSeuratObject(counts = inputMatrix, min.cells = 3, min.features = 20)
dta <- check_cluster(dta)
}
