% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Seurat.R
\name{check_Seurat}
\alias{check_Seurat}
\title{Check whether the object is Seurat object or not}
\usage{
check_Seurat(obj)
}
\arguments{
\item{obj}{The object to check whether it is Seurat object or not}
}
\value{
An error message if the objective is not a Seurat objective and "TRUE" if it is a Seurat objective
}
\description{
Check whether the object is Seurat object or not
}
