% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doClustering.R
\name{doClustering}
\alias{doClustering}
\title{Cluster Determination}
\usage{
doClustering(
  obj,
  datatype = "RNA",
  cluster_cell = NULL,
  dims = 1:50,
  k.param = 30,
  resolution = 0.5,
  hclustmethod = "complete",
  ncluster = 3
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{datatype}{Data type for your data, default is 'datatype = "RNA"', which is used for scRNAseq data.}

\item{cluster_cell}{The cluster result for cells if it is already known.}

\item{dims}{An integer value. Define dimensions of reduction to use as input. (Do cluster for single cell data.)}

\item{k.param}{An integer value. Defines k for the k-nearest neighbor algorithm. (Do cluster for single cell data.)}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities. (Do cluster for single cell data.)}

\item{hclustmethod}{The agglomeration method to be used for hierarchical clustering, defalut is "complete".}

\item{ncluster}{An integer, which is the number of cluster when your input including results from hierarchical clustering.}
}
\value{
It will return a Seurat object with cluster.
}
\description{
It will do clustering. If the data type is single cell data, the input must be Seurat object and it will use the “Findcluster” function in the Seurat package
For any other data type, it will do hierarchical clustering.
}
\examples{
pbmc_example <- scqc(small_RNA, min.cells = 1, min.features = 10, nfeatures = 100, npcs = 10)
pbmc_example <- doClustering(pbmc_example, dims = 1:10, k.param = 5, resolution = 0.75)
head(pbmc_example@meta.data$seurat_clusters)
}
