% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_analysis.R
\name{GSEA_analysis}
\alias{GSEA_analysis}
\title{Title Runs preranked gene set enrichment analysis.}
\usage{
GSEA_analysis(obj, cluster_list, db = "PanglaoDB_list", minSize = 15,
  maxSize = 500, nperm = 10000)
}
\arguments{
\item{obj}{A Seurat object with cluster.}

\item{cluster_list}{A ranked gene list for each cluster.}

\item{db}{The cell type data base to use. It should be 'PanglaoDB_list' for 'PanglaoDB' data base and 'GSEA_list' for 'GSEA' data base.}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{nperm}{Number of permutations to do. Minimial possible nominal p-value is about 1/nperm}
}
\value{
Cell type for each cluster.
}
\description{
The function do gene set enrichment analysis(GSEA) for each cluster with its gene ranks. It will return cell type for each cluster.
}
