% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSEA_list}
\alias{GSEA_list}
\title{Cell type gene expression markers from GSEA data base}
\format{
A list with 257 cell types
}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/supplementary_genesets.jsp#SCSig}
}
\usage{
GSEA_list
}
\description{
A list contains cell types with their marker gene
}
\keyword{datasets}
