% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelCelltype.R
\name{labelCelltype}
\alias{labelCelltype}
\title{Add cell type to Seurat object.}
\usage{
labelCelltype(obj, cluster_celltype)
}
\arguments{
\item{obj}{A Seurat object with cluster.}

\item{cluster_celltype}{A matrix. Each row is one cell type, contains NSF and padj.}
}
\value{
It will return a list including a Seurat object with cell type for each cell and a cell type matrix.
}
\description{
Add cell type to Seurat object.
}
\examples{
pbmc_example <- scqc(small_RNA, min.cells = 1, min.features = 10, nfeatures = 100, npcs = 10)
pbmc_example <- doClustering(pbmc_example, dims = 1:10, k.param = 5, resolution = 0.75)
cluster_list <- getFC(pbmc_example, min.pct = 0.25, test.use = "MAST")
cluster_celltype <- doGSEA(cluster_list = cluster_list, minSize = 5)
celltype <- labelCelltype(pbmc_example, cluster_celltype)

}
