% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelSeurat.R
\name{labelSeurat}
\alias{labelSeurat}
\title{Add cell type to Seurat object}
\usage{
labelSeurat(obj, cluster_celltype)
}
\arguments{
\item{obj}{A Seurat object with cluster.}

\item{cluster_celltype}{A matrix. Each row is one cell type, contains NSF and padj.}
}
\value{
A list contains two entries. The first is a Seurat object with cell type for each cell and the second is a cell matrix.
}
\description{
Add cell type to Seurat object
}
\examples{
pbmc_example <- scqc(pbmc_small, min.cells = 1, min.features = 10)
pbmc_example <- doClustering(pbmc_example, nfeatures = 100, npcs = 10,
                              dims = 1:10, k.param = 5, resolution = 0.75)
cluster_list <- getFC(pbmc_example, min.pct = 0.25, test.use = 'MAST')
cluster_celltype <- doGSEA(cluster_list = cluster_list, minSize = 5)
pbmc_example <- labelSeurat(pbmc_example, cluster_celltype)$obj

}
