% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PanglaoDB_list}
\alias{PanglaoDB_list}
\title{Cell type gene expression markers from GSEA data base}
\format{
A list with 178 cell types
}
\source{
\url{https://panglaodb.se/markers.html}
}
\usage{
PanglaoDB_list
}
\description{
A list contains cell types with their marker gene
}
\keyword{datasets}
