% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_raw}
\alias{pbmc_raw}
\title{PBMC raw data set.}
\format{A txt file with 32738 rows and 2700 columns}
\usage{
pbmc_raw
}
\description{
A version of 10X Genomics' 3k PBMC dataset
}
\keyword{datasets}
