% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scqc.R
\name{scqc}
\alias{scqc}
\title{Single cell quality control.}
\usage{
scqc(obj, min.cells = 10, min.features = 1000, percent.mt = 5, oversd = NULL)
}
\arguments{
\item{obj}{A seurat object or a raw count gene expression.}

\item{min.cells}{An integer value. Include features detected in at least this many cells.}

\item{min.features}{An integer value. Include cells where at least this many features are detected.}

\item{percent.mt}{A decimal value between 0 and 1. Define the highest percentage of reads that map to the mitochondrial genome.}

\item{oversd}{Remove cells whose library size is greater than mean + oversd * sd. Default is null, which doesn't remove cells.}
}
\value{
If the input is not Seurat object, it will assume it is a count gene expression matrix and transform it as a seurat object. And then do quality control to select cells. Otherwise, it will do quality control directly.
}
\description{
Check whether the input is a seurat object or not and then perform single-cell data quality control, including checking for minimum cell library size, mitochondrial ratio, outlier cells, and the fraction of cells where a gene is expressed.
}
\examples{
pbmc_example <- scqc(pbmc_small, min.cells = 1, min.features = 10)
pbmc_example
}
