% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTypeGSEA.R
\name{scTypeGSEA}
\alias{scTypeGSEA}
\title{Gene set enrichment analysis for single cell data.}
\usage{
scTypeGSEA(obj, ...)
}
\arguments{
\item{obj}{An Seurat object.}

\item{...}{Arguments passed to other methods}
}
\value{
An Seurat object with celltype for each cluster.
}
\description{
Gene set enrichment analysis for single cell data.
}
