% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_pbmc_rna}
\alias{small_pbmc_rna}
\title{A small example version of the PBMC dataset}
\format{
A matrix with 230 rows and 80 columns
}
\source{
\url{https://github.com/satijalab/seurat/blob/master/inst/extdata/pbmc_raw.txt}
}
\usage{
small_pbmc_rna
}
\description{
A subsetted version of 10X Genomics' 3k PBMC dataset
}
\keyword{datasets}
