% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAIR_list}
\alias{TAIR_list}
\title{The reference genome for Arabidopsis.}
\format{
A list with 16 cell types
}
\usage{
TAIR_list
}
\description{
A list contains cell types with their marker gene
}
\keyword{datasets}
