% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_DE_cluster.R
\name{Test_DE_cluster}
\alias{Test_DE_cluster}
\title{Title Gene rank of identity cluster}
\usage{
Test_DE_cluster(obj, min.pct = 0.25, test.use = "wilcox")
}
\arguments{
\item{obj}{A Seurat object with cluster.}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed.}

\item{test.use}{Denotes which test to use. Available options are 'wilcox', 'bimod', 'roc', 'negbinom', 'poisson', 'LR', 'MAST' and 'DESeq2'.}
}
\value{
For each cluster, there will be a ranked gene list.
}
\description{
This function is to do DE for each cluster regarding all other clusters.
And then for each cluster, it will return a ranked gene list.
}
