% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seurat.R
\name{check_seurat}
\alias{check_seurat}
\title{Title Check Seurat.}
\usage{
check_seurat(obj, min.cells = 3, min.features = 200)
}
\arguments{
\item{obj}{A seurat object or a count gene expression.}

\item{min.cells}{Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are detected.}
}
\value{
If the input is a seurat object, just return it. If not, it will assume it is a count gene expression matrix and transform it as a seurat object.
}
\description{
Check whether the input is a seurat object or not.
}
\examples{
pbmc_example <- check_seurat(pbmc_raw, min.cells = 1, min.features = 10)
}
