% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doClustering.R
\name{doClustering}
\alias{doClustering}
\title{Cluster Determination}
\usage{
doClustering(
  obj,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  selection.method = "vst",
  nfeatures = 2000,
  npcs = 50,
  dims = 1:50,
  k.param = 30,
  resolution = 0.5,
  doit = "FALSE",
  doprocess = "FALSE"
)
}
\arguments{
\item{obj}{An Seurat object.}

\item{normalization.method}{Method for normalization. Include 'LogNormalize', 'CLR' and 'RC'.}

\item{scale.factor}{Sets the scale factor for cell-level normalization.}

\item{selection.method}{How to choose top variable features. Include 'vst', 'mean.var.plot' and 'dispersion'.}

\item{nfeatures}{An integer value. Define the number of features to select as top variable features.}

\item{npcs}{An integer value. Define total Number of PCs to compute and store (50 by default).}

\item{dims}{An integer value. Define dimensions of reduction to use as input. (For cluster.)}

\item{k.param}{An integer value. Defines k for the k-nearest neighbor algorithm. (For cluster.)}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities. (For cluster.)}

\item{doit}{A boolean value (TRUE/FALSE). If true, the function will do the cluster with new parameters even the object contains original cluster. (Default is FALSE.)}

\item{doprocess}{A boolean value (TRUE/FALSE). If true, the function will do data process with new parameters. (Default is FALSE.)}
}
\value{
If the input is already clustered, just return it. If not, the function do cluster and return obj with cluster.
}
\description{
If the input is already clustered, just return it.
If not, the function will check whether the data has already been pre-processed, for example, PCA. If not, the function will do data pre-process first.
After the basic data pre-process (Normalization, Scale data, Find HVG and PCA), the function will do cluster.
}
\examples{
pbmc_example <- scqc(pbmc_small, min.cells = 1, min.features = 10)
is.null(pbmc_example@meta.data$seurat_clusters)
pbmc_example <- doClustering(pbmc_example, nfeatures = 100, npcs = 10,
                              dims = 1:10, k.param = 5, resolution = 0.5)
head(pbmc_example@meta.data$seurat_clusters)
pbmc_example <- doClustering(pbmc_example, nfeatures = 100, npcs = 10,
                              dims = 1:10, k.param = 5, resolution = 0.75, doit = TRUE)
head(pbmc_example@meta.data$seurat_clusters)
}
