% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_raw}
\alias{pbmc_raw}
\title{A small example version of the PBMC dataset}
\format{A txt file with 230 rows and 80 columns}
\source{
\url{https://github.com/satijalab/seurat/blob/master/inst/extdata/pbmc_raw.txt}
}
\usage{
pbmc_raw
}
\description{
A subsetted version of 10X Genomics' 3k PBMC dataset
}
\keyword{datasets}
