% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scqc.R
\name{scqc}
\alias{scqc}
\title{Single cell quality control and basic data pre-process.}
\usage{
scqc(
  obj,
  datatype = "RNA",
  min.cells = 10,
  min.features = 1000,
  percent.mt = 10,
  oversd = NULL,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  selection.method = "vst",
  nfeatures = 2000,
  npcs = 50
)
}
\arguments{
\item{obj}{A seurat object or a raw count gene expression.}

\item{datatype}{Data type for your data, default is 'datatype = "RNA"', which is used for scRNAseq data.}

\item{min.cells}{An integer value. Include features detected in at least this many cells.}

\item{min.features}{An integer value. Include cells where at least this many features are detected.}

\item{percent.mt}{Define the highest percentage of reads that map to the mitochondrial genome.}

\item{oversd}{Remove cells whose library size is greater than mean + oversd * sd. Default is null, which doesn't remove cells.}

\item{normalization.method}{Method for normalization. Include 'LogNormalize', 'CLR' and 'RC'.}

\item{scale.factor}{Sets the scale factor for cell-level normalization.}

\item{selection.method}{How to choose top variable features. Include 'vst', 'mean.var.plot' and 'dispersion'.}

\item{nfeatures}{An integer value. Define the number of features to select as top variable features.}

\item{npcs}{An integer value. Define total Number of PCs to compute and store (50 by default).}
}
\value{
A Seurat Object with quality control and basic data pre-process.
}
\description{
For single cell type data, it will create a Seurat object and perform single-cell data quality control,
including checking for minimum cell library size, mitochondrial ratio, outlier cells, and the fraction of cells where a gene is expressed.
And then the function will do basic data pre-process, which includes “Normalization”, “Scale data”, “Find HVG” and “PCA”.
For other type of data, it will create a Seurat object to save the matrix.
}
\examples{
pbmc_example <- scqc(small_pbmc_rna, min.cells = 1, min.features = 10, nfeatures = 100, npcs = 10)
pbmc_example
}
